possible fix for Drive_IO Error, issue #5
Note: we only do this for floppy disks to try and avoid sharing issue that caused this to be commented out in the first place.  The condition may need further refining to avoid sharing violation, but not sure what conditions actually caused the sharing violation to occur.  Or we may need to do a different call to force kernel to reload its idea of what the drive geometry is.