Initial French & Turkish translations
Thanks a lot for making format translatable, this will also help during the install process. These initial translations could be useful for testing purposes. 
Also note that English is usually a very concise language compared to French for instance; so perhaps it would be helpful to add a comment to the file explaining how to add extra keys.